+++ src/generator/gen_json.rs
@@ -93,9 +93,11 @@ impl GeneratorSw for GeneratorJson {
             }
             self.write(&format!("{tab}   ],\n"));
         }
+        // Optional number of fractional bits
         if field.nb_frac!=0 {
             self.write(&format!("{tab}   \"nb_frac\" : {},\n", field.nb_frac));
         }
+        // Add field description
         self.write(&format!("{tab}   \"desc\" : \"{desc}\"\n"));
         let sep = if is_last {""} else {","};
         self.write( &format!("{tab}}}{sep}\n"));